refactor

+++ payment-service/src/main/java/com/modsen/paymentservice/dto/BalanceResponse.java
@@ -5,10 +5,12 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.experimental.FieldDefaults;
 
+import java.math.BigDecimal;
+
 @Builder
 @Getter
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class BalanceResponse {
-    long amount;
+    BigDecimal amount;
     String currency;
 }
