+++ payment-service/src/main/java/com/modsen/paymentservice/dto/ChargeResponse.java
@@ -5,11 +5,13 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.experimental.FieldDefaults;
 
+import java.math.BigDecimal;
+
 @FieldDefaults(level = AccessLevel.PRIVATE)
 @Getter
 @Builder
 public class ChargeResponse {
     String id;
     String currency;
-    long amount;
+    BigDecimal amount;
 }
